arc/requirements.txt
allure-behave==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==2.2.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
behave==1.2.6
behave-pytest==0.1.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
docopt==0.6.2
et-xmlfile==1.1.0
Faker==13.12.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-resources==5.7.1
jdcal==1.4.1
Jinja2==3.1.2
jsonschema==4.5.1
junit2html==30.1.3
lxml==4.8.0
MarkupSafe==2.1.1
numpy==1.22.4
oauthlib==3.2.0
openpyxl==2.4.11
outcome==1.1.0
pandas==1.4.2
parse==1.19.0
parse-type==0.6.0
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyrsistent==0.18.1
PySocks==1.7.1
pytest==2.9.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
pywin32==302
PyYAML==6.0
records==0.5.3
requests==2.27.1
requests-oauthlib==1.3.1
screeninfo==0.8
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
tablib==3.2.1
teamcity-messages==1.31
termcolor==1.1.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
zipp==3.8.0
